[Refactor] 주석 추가
