+++ Assets/01.Scenes/JungHa/Earthquake.unity
@@ -8010,7 +8010,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &470329511
 RectTransform:
   m_ObjectHideFlags: 0
@@ -8024,7 +8024,7 @@ RectTransform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1148326296}
-  m_Father: {fileID: 827983244}
+  m_Father: {fileID: 828131336}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -13469,7 +13469,6 @@ RectTransform:
   m_LocalScale: {x: 0, y: 0, z: 0}
   m_ConstrainProportionsScale: 0
   m_Children:
-  - {fileID: 1469372851}
   - {fileID: 1280476632}
   - {fileID: 726807437}
   - {fileID: 1697663662}
@@ -14933,7 +14932,6 @@ Transform:
   - {fileID: 758703205}
   - {fileID: 466543192}
   - {fileID: 1048204300}
-  - {fileID: 470329511}
   m_Father: {fileID: 828131336}
   m_LocalEulerAnglesHint: {x: 0, y: 90, z: 0}
 --- !u!64 &827983245
@@ -15027,6 +15025,7 @@ Transform:
   m_Children:
   - {fileID: 1567413672}
   - {fileID: 827983244}
+  - {fileID: 470329511}
   m_Father: {fileID: 281172094}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &845115067 stripped
@@ -18769,7 +18768,7 @@ PrefabInstance:
     - target: {fileID: 2741514586999573498, guid: 894edc677057b42b28be567d8fa86b21,
         type: 3}
       propertyPath: m_text
-      value: "\uC2E4\uC804 \uBAA8\uB4DC\uAC00 \uC885\uB8CC\uB418\uC5C8\uC2B5\uB2C8\uB2E4."
+      value: "\uC2E4\uC804 \uBAA8\uB4DC\uAC00 \uC885\uB8CC\uB418\uC5C8\uC2B5\uB2C8\uB2E4.\n\n<\uD53C\uB4DC\uBC31>"
       objectReference: {fileID: 0}
     - target: {fileID: 2741514586999573498, guid: 894edc677057b42b28be567d8fa86b21,
         type: 3}
@@ -18920,7 +18919,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   hasInteracted: 0
   interactIndex: -1
-  UI_elevator: {fileID: 470329510}
+  elevatorUI: {fileID: 1148326299}
 --- !u!1001 &1050465404
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -20805,13 +20804,14 @@ GameObject:
   - component: {fileID: 1148326296}
   - component: {fileID: 1148326298}
   - component: {fileID: 1148326297}
+  - component: {fileID: 1148326299}
   m_Layer: 5
   m_Name: TextPannel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!224 &1148326296
 RectTransform:
   m_ObjectHideFlags: 0
@@ -20870,6 +20870,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1148326295}
   m_CullTransparentMesh: 1
+--- !u!114 &1148326299
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1148326295}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 743ddb64ca4054c49b916cec9e62c5fa, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1001 &1150447812
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -23826,7 +23838,7 @@ GameObject:
   - component: {fileID: 1280476634}
   - component: {fileID: 1280476633}
   m_Layer: 5
-  m_Name: StartTimerText
+  m_Name: StartText
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -23848,8 +23860,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 14}
-  m_SizeDelta: {x: 300, y: 100}
+  m_AnchoredPosition: {x: 0.0025024, y: 0.000030518}
+  m_SizeDelta: {x: 819.51, y: 529}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1280476633
 MonoBehaviour:
@@ -23873,10 +23885,10 @@ MonoBehaviour:
       m_Calls: []
   m_FontData:
     m_Font: {fileID: 10102, guid: 0000000000000000e000000000000000, type: 0}
-    m_FontSize: 40
+    m_FontSize: 27
     m_FontStyle: 0
     m_BestFit: 0
-    m_MinSize: 4
+    m_MinSize: 0
     m_MaxSize: 40
     m_Alignment: 4
     m_AlignByGeometry: 0
@@ -23884,7 +23896,10 @@ MonoBehaviour:
     m_HorizontalOverflow: 0
     m_VerticalOverflow: 0
     m_LineSpacing: 1
-  m_Text: "\uC2E4\uC804\uBAA8\uB4DC \uC2DC\uC791"
+  m_Text: "\uC5F0\uC2B5 \uBAA8\uB4DC\uC5D0\uC11C \uBC30\uC6B4 \uB0B4\uC6A9\uC744
+    \uB5A0\uC62C\uB824 \n\uC9C0\uC9C4\uC774 \uB098\uB294 \uC0C1\uD669\uC5D0\uC11C
+    \uC548\uC804\uD558\uAC8C \uC544\uD30C\uD2B8 \uBC16\uC73C\uB85C \uB300\uD53C\uD574\uBCF4\uC138\uC694!\n(\uC784\uC2DC
+    \uBA54\uC138\uC9C0)"
 --- !u!222 &1280476634
 CanvasRenderer:
   m_ObjectHideFlags: 0
@@ -27432,85 +27447,6 @@ Light:
   m_UseViewFrustumForShadowCasterCull: 1
   m_ShadowRadius: 0
   m_ShadowAngle: 0
---- !u!1 &1469372850
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 1469372851}
-  - component: {fileID: 1469372853}
-  - component: {fileID: 1469372852}
-  m_Layer: 5
-  m_Name: StartMessage
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!224 &1469372851
-RectTransform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1469372850}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
-  m_ConstrainProportionsScale: 0
-  m_Children: []
-  m_Father: {fileID: 735108051}
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: 13.999999}
-  m_SizeDelta: {x: 500, y: 100}
-  m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &1469372852
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1469372850}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5f7201a12d95ffc409449d95f23cf332, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_Material: {fileID: 0}
-  m_Color: {r: 0.19607843, g: 0.19607843, b: 0.19607843, a: 1}
-  m_RaycastTarget: 1
-  m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
-  m_Maskable: 1
-  m_OnCullStateChanged:
-    m_PersistentCalls:
-      m_Calls: []
-  m_FontData:
-    m_Font: {fileID: 10102, guid: 0000000000000000e000000000000000, type: 0}
-    m_FontSize: 60
-    m_FontStyle: 0
-    m_BestFit: 0
-    m_MinSize: 6
-    m_MaxSize: 60
-    m_Alignment: 4
-    m_AlignByGeometry: 0
-    m_RichText: 1
-    m_HorizontalOverflow: 0
-    m_VerticalOverflow: 0
-    m_LineSpacing: 1
-  m_Text: Game Start!
---- !u!222 &1469372853
-CanvasRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1469372850}
-  m_CullTransparentMesh: 1
 --- !u!1001 &1472407639
 PrefabInstance:
   m_ObjectHideFlags: 0
