+++ Assets/02.Scripts/EarthquakeScript/ElevatorUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 743ddb64ca4054c49b916cec9e62c5fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
